+++ config/local.ts
@@ -1,22 +1,19 @@
-import {Config} from './Config';
+import {Config} from './types';
+import {readValidators} from './utils';
 
 const local: Config = {
   contractRegistry: {
     address: '0x7c2C195CD6D34B8F845992d380aADB2730bB9C6F'
   },
   token: {
-    totalSupply: '1000000000000000000',
+    totalSupply: '1000' + '0'.repeat(18),
     name: 'Umbrella',
     symbol: 'UMB'
   },
   chain: {
     blockPadding: 6
   },
-  validators: [
-    {
-      location: 'http://localhost:3000'
-    }
-  ]
+  validators: readValidators('http://localhost:3000'),
 };
 
 export default local;