+++ contracts/ValidatorRegistry.sol
@@ -16,13 +16,9 @@ contract ValidatorRegistry is IValidatorRegistry, Ownable {
 
   address[] override public addresses;
 
-  event LogValidatorRegistered(
-    address id
-  );
-
-  event LogValidatorUpdated(
-    address id
-  );
+  event LogValidatorRegistered(address id);
+  event LogValidatorUpdated(address id);
+  event LogValidatorRemoved(address id);
 
   function create(address _id, string calldata _location) override external onlyOwner {
     Validator storage validator = validators[_id];
@@ -37,6 +33,21 @@ contract ValidatorRegistry is IValidatorRegistry, Ownable {
     LogValidatorRegistered(validator.id);
   }
 
+  function remove(address _id) external onlyOwner {
+    require(validators[_id].id != address(0x0), "validator NOT exists");
+
+    delete validators[_id];
+    emit LogValidatorRemoved(_id);
+
+    for (uint256 i = 0; i < addresses.length; i++) {
+      if (addresses[i] == _id) {
+        addresses[i] = addresses[addresses.length - 1];
+        addresses.pop();
+        return;
+      }
+    }
+  }
+
   function update(address _id, string calldata _location) override external onlyOwner {
     Validator storage validator = validators[_id];
 
@@ -49,5 +60,9 @@ contract ValidatorRegistry is IValidatorRegistry, Ownable {
 
   function getNumberOfValidators() override external view returns (uint256) {
     return addresses.length;
+  }
+
+  function getName() override external pure returns (bytes32) {
+    return "ValidatorRegistry";
   }
 }