+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+## [0.3.0] - 2021-03-15
+### Added
+- option to remove validator + update scripts
+
+### Fixed
+- cycle leader selection
+
 ## [0.2.0] - 2021-03-10
 ### Added
 - individual script for deploy/update `Chain`